refactor: android example app
