+++ example/android/app/src/main/java/sourcepoint/reactnativecmp/example/MainActivity.kt
@@ -1,18 +1,17 @@
-package com.reactnativecmpexample
+package sourcepoint.reactnativecmp.example
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
 
-class
-MainActivity : ReactActivity() {
+class MainActivity : ReactActivity() {
 
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "SourcepointCmpExample"
+  override fun getMainComponentName(): String = "ReactNativeCmpExample"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
