+++ example/android/settings.gradle
@@ -1,4 +1,6 @@
-rootProject.name = 'SourcepointCmpExample'
-apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
+pluginManagement { includeBuild("../node_modules/@react-native/gradle-plugin") }
+plugins { id("com.facebook.react.settings") }
+extensions.configure(com.facebook.react.ReactSettingsExtension){ ex -> ex.autolinkLibrariesFromCommand() }
+rootProject.name = 'sourcepoint.reactnativecmp.example'
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
